backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
beautifulsoup4
langchain
chromadb
sentence-transformers
groq
playwright
pathlib
pypdf
ipykernel
pandas
pyarrow
fastparquet
datasets   
langchain_text_splitters
faiss-cpu
hf_xet
langchain_groq
pymongo
pylance
twilio
python-multipart